+++ backend/NuGetLibraries/Cinedex.WebService.Contracts/Requests/ForgotPasswordRequest.cs
@@ -1,0 +1,6 @@
+namespace Cinedex.WebService.Contracts.Requests;
+
+public class ForgotPasswordRequest
+{
+    public required string Email { get; init; } = string.Empty;
+}