+++ backend/NuGetLibraries/Cinedex.WebService.Contracts/Requests/ResetPasswordRequest.cs
@@ -1,0 +1,10 @@
+namespace Cinedex.WebService.Contracts.Requests;
+
+public class ResetPasswordRequest
+{
+    public required string Email { get; init; } = string.Empty;
+
+    public required string ResetToken { get; init; } = string.Empty;
+
+    public required string NewPassword { get; init; } = string.Empty;
+}